requirements.txt
# Hey Nova MVP Dependencies
# Core audio and speech processing
faster-whisper>=0.9.0
pyaudio>=0.2.11
numpy>=1.24.0

# Wake word detection
pvporcupine>=2.2.0

# OpenAI integration
openai>=1.0.0

# Timezone and utilities
pytz>=2023.3

# HTTP requests for APIs
requests>=2.31.0

# Notion API integration
notion-client>=2.0.0

# Google Calendar API integration
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0

# Voice Activity Detection for natural conversations
webrtcvad>=2.0.10

# Audio processing
sounddevice>=0.4.6

# Future dependencies (commented out for MVP)
# fastapi>=0.104.0
# uvicorn>=0.24.0
# websockets>=12.0
# sqlite3  # Built into Python
# cryptography>=41.0.0
